backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
PyMySQL==1.1.0
cryptography==41.0.7
bcrypt==4.1.2
python-dotenv==1.0.0
Werkzeug==3.0.1
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
reportlab==4.0.7
requests==2.31.0
colorama==0.4.6
